fix compilation error

+++ examples/Test/Test.ino
@@ -13,7 +13,7 @@ DeviceDriver* myDriver;
 address myAddr = 1;
 
 void setup() {
-    myDriver = new EbyteDeviceDriver(LORA_RX, LORA_TX, LORA_M0, LORA_M1, LORA_AUX, myAddr, 9);
+    myDriver = new EbyteDeviceDriver(LORA_RX, LORA_TX, LORA_M0, LORA_M1, LORA_AUX, myAddr, (uint8_t)9);
     //myEngine = new ForwardEngine(myAddr,myDriver);
     myDriver->init();
